Do strength reduction on write loops
Apparently ghc/llvm do better with build/fold style loops for indexing
arrays. However applyPreemptives has to do its own control flow so for
now we only do it for the loops that apply bit masks.

The payoff isn't huge - 3.8 to 3.5 seconds when running single threaded
- but it isn't much uglier than the old version. We could do the same
for hiddenSingletonPass and we could do it for Preemptive pass if we
implement searchPreemptives on top of permutations instead of a
recursive depth first search. This would make the loop tail recursive
but makes pruning much harder so whether this is an improvement likely
depends on the inputs.